pyproject.toml
[tool.poetry]
name = "opinion_analyzer"
version = "0.1.0"
description = "Code for the Innocheck Project SmartOpinion"
authors = ["Veton Matoshi"]
license = "Not specified yet."

[tool.poetry.dependencies]
python = "^3.10"
beautifulsoup4 = "4.12.3"
fasttext = "0.9.3"
minio = "^7.2.7"
nltk = "^3.8.1"
numpy = "1.26.4"
openai = "^1.34.0"
pandas = "^2.2.2"
peft = "^0.11.1"
pillow = "^10.3.0"
# pydantic = "^2.7.4"
PyPDF2 = "^3.0.1"
PyYAML = "^6.0.1"
rapidfuzz = "^3.9.3"
ray = "^2.24.0"
requests = "^2.32.3"
rouge-score = "^0.1.2"
scikit-learn = "^1.5.0"
sentence-transformers = "^3.0.1"
setfit = "^1.0.3"
spacy = "^3.7.0"
torch = "^2.1.2"
tqdm = "^4.66.4"
transformers = "4.43.1"
tabula-py = "^2.9.3"
datasets = "^2.20.0"
bitsandbytes = "^0.43.1"
accelerate = "0.33.0"
xformers = "^0.0.27"
openpyxl = "3.1.5"
pymupdf = "^1.24.5"
optuna = "^3.6.1"
autoawq = "^0.2.5"
sentencepiece = "^0.2.0"
vllm = "^0.5.3.post1"
huggingface-hub = "0.23.2"
pytest = "^8.3.2"
pdf2image= "1.17.0"
dateparser = "1.2.0"
setproctitle = "1.3.3"
swissparlpy = "0.3.0"
dspy-ai = "2.4.15"
textgrad = "0.1.4"
tensorboard = "2.18.0"
wandb = "^0.18.3"
[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pylint = "^3.2.3"
black = "^24.4.2"
taskipy = "^1.13.0"

[tool.taskipy.tasks]
setup = "poetry run python setup_everything.py"
